fix: Refactored theme service into Navigation.Toolkit because of toolkit dependency
