+++ samples/Playground/Playground.Shared/Views/ThemeSwitchPage.xaml.cs
@@ -1,17 +1,21 @@
-﻿using Uno.Extensions.Navigation.UI;
+﻿using Uno.Extensions.Navigation.Toolkit.UI;
+using Uno.Extensions.Navigation.UI;
 
 // The Blank Page item template is documented at https://go.microsoft.com/fwlink/?LinkId=234238
 
 namespace Playground.Views
 {
-	public sealed partial class ThemeSwitchPage : Page
+	public sealed partial class ThemeSwitchPage : Page, IInjectable<IServiceProvider>
 	{
+
+		// NR: Move this functionality into a ViewModel and use data binding.
+		// The IThemeService can be injected into the constructor of the ViewModel
+
+
 		private IThemeService ts;
 		public ThemeSwitchPage()
 		{
 			this.InitializeComponent();
-			ts = (Application.Current as App)?.ThemeService;
-			ts.DesiredThemeChanged += ts_DesiredThemeChanged;
 		}
 
 		private void ts_DesiredThemeChanged(object? sender, DesiredTheme e)
@@ -33,6 +37,12 @@ namespace Playground.Views
 		private async void ButtonLight_Click(object sender, RoutedEventArgs e)
 		{
 			await ts.SetThemeAsync(DesiredTheme.Light);
+		}
+
+		public void Inject(IServiceProvider entity)
+		{
+			ts = entity.GetRequiredService<IThemeService>();
+			ts.DesiredThemeChanged += ts_DesiredThemeChanged;
 		}
 	}
 }
