+++ src/Uno.Extensions.Navigation.Toolkit/DesiredTheme.cs
@@ -1,0 +1,8 @@
+﻿namespace Uno.Extensions.Navigation.UI;
+
+public enum DesiredTheme
+{
+	System = default,
+	Light = 1,
+	Dark = 2
+}
