+++ src/Uno.Extensions.Navigation.Toolkit/GlobalUsings.cs
@@ -9,6 +9,7 @@ global using Microsoft.Extensions.Hosting;
 global using Microsoft.Extensions.Logging;
 global using Windows.Foundation;
 global using Windows.UI.Popups;
+global using Uno.Extensions.Configuration;
 global using Uno.Extensions.Hosting;
 global using Uno.Extensions.Navigation;
 global using Uno.Extensions.Navigation.Navigators;
