+++ src/Uno.Extensions.Navigation.Toolkit/IThemeService.cs
@@ -1,0 +1,24 @@
+﻿namespace Uno.Extensions.Navigation.Toolkit.UI;
+
+public interface IThemeService
+{
+	/// <summary>
+	/// Get if the application is currently in dark mode.
+	/// </summary>
+	bool IsDark { get; }
+
+	/// <summary>
+	///  Get the previously saved theme.
+	/// </summary>
+	DesiredTheme Theme { get; }
+
+	/// <summary>
+	/// Sets the system theme for the provided XamlRoot.
+	/// </summary>
+	Task SetThemeAsync(DesiredTheme theme = DesiredTheme.System);
+
+	/// <summary>
+	/// Event that fires up whenever user updates current theme.
+	/// </summary>
+	event EventHandler<DesiredTheme> DesiredThemeChanged;
+}
