+++ src/Uno.Extensions.Navigation.Toolkit/ThemeSettings.cs
@@ -1,0 +1,6 @@
+﻿namespace Uno.Extensions.Navigation.UI;
+
+public record ThemeSettings
+{
+	public DesiredTheme CurrentTheme { get; set; }
+}
