--- src/Uno.Extensions.Navigation.UI/DesiredTheme.cs
@@ -1,13 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace Uno.Extensions.Navigation.UI
-{
-	public enum DesiredTheme
-	{
-		System = default,
-		Light = 1,
-		Dark = 2
-	}
-}
