--- src/Uno.Extensions.Navigation.UI/ThemeSettings.cs
@@ -1,11 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace Uno.Extensions.Navigation.UI
-{
-	public record ThemeSettings
-	{
-		public DesiredTheme CurrentTheme { get; set; }
-	}
-}
