+++ src/Uno.Extensions.Navigation.UI/Uno.Extensions.Navigation.UI.csproj
@@ -12,7 +12,6 @@
 	</PropertyGroup>
 
 	<ItemGroup>
-		<PackageReference Include="Uno.Toolkit.UI" />
 		<PackageReference Include="Uno.UI" />
 	</ItemGroup>
 
@@ -22,10 +21,6 @@
 
 	<ItemGroup>
 	  <ProjectReference Include="..\Uno.Extensions.Hosting.UI\Uno.Extensions.Hosting.UWP.csproj" />
-	</ItemGroup>
-
-	<ItemGroup>
-	  <ProjectReference Include="..\Uno.Extensions.Configuration\Uno.Extensions.Configuration.csproj" />
 	</ItemGroup>
 	
 </Project>
