+++ src/Uno.Extensions.Navigation.UI/Uno.Extensions.Navigation.WinUI.csproj
@@ -20,12 +20,4 @@
 	<ItemGroup>
 	  <ProjectReference Include="..\Uno.Extensions.Hosting.UI\Uno.Extensions.Hosting.WinUI.csproj" />
 	</ItemGroup>
-
-	<ItemGroup>
-	  <PackageReference Include="Uno.Toolkit.WinUI" />
-	</ItemGroup>
-
-	<ItemGroup>
-	  <ProjectReference Include="..\Uno.Extensions.Configuration\Uno.Extensions.Configuration.csproj" />
-	</ItemGroup>
 </Project>
